Moved code from `dev` package into separate packages. Removed dead code.
