+++ lumy_middleware/kiara_modules/__init__.py
@@ -1,4 +1,4 @@
 from kiara import (KiaraEntryPointItem, find_kiara_modules_under)
 
 modules: KiaraEntryPointItem = (find_kiara_modules_under, [
-                                "lumy_middleware.dev.modules"])
+                                "lumy_middleware.kiara_modules"])
