+++ lumy_middleware/kiara_modules/network_analysis_temporary.py
@@ -5,9 +5,74 @@ import pandas as pd
 import pyarrow as pa
 from kiara import Kiara
 from kiara.data.values import ValueSchema
-from kiara.module import KiaraModule, StepInputs, StepOutputs
+from kiara.module import KiaraModule, ValueSet
+from lumy_middleware.context.kiara.dataregistry import get_value_label
 
 logger = logging.getLogger(__name__)
+
+
+class DataSelectionModule(KiaraModule):
+
+    def create_input_schema(self) -> Mapping[str, ValueSchema]:
+        return {
+            "selectedItemsIds": ValueSchema(
+                type="any", doc="URIs of selected items.", default=[]
+            ),
+            "metadataFields": ValueSchema(
+                type="any", doc="Metadata field.", default=[]
+            )
+        }
+
+    def create_output_schema(self) -> Mapping[str, ValueSchema]:
+        return {
+            "selectedItems": ValueSchema(
+                type="table",
+                doc="Selected corpus.",
+            )
+        }
+
+    def process(self, inputs: ValueSet, outputs: ValueSet) -> None:
+        selected_items_ids = inputs.get_value_data(
+            'selectedItemsIds') or []
+
+        def get_value(id: str):
+            '''
+            TODO: This will go away when workflow is refactored.
+            registry should not be accessible from modules code.
+            '''
+            return self._kiara.data_store.get_value_obj(id)
+
+        fields = ['id', 'label', 'type', 'columnNames', 'columnTypes']
+        selected_items = [
+            {
+                'id': id,
+                'label': get_value_label(get_value(id), self._kiara),
+                'type': get_value(id).type_name,
+                'columnNames': get_value(id)
+                .get_metadata('table').get('table', {})
+                .get('column_names', []),
+                'columnTypes': [
+                    v['arrow_type_name']
+                    for v in get_value(id)
+                    .get_metadata('table').get('table', {}).get('schema', {})
+                    .values()
+                ]
+            }
+            for id in selected_items_ids
+        ]
+
+        selected_items = pa.Table.from_pydict({
+            k: [i[k] for i in selected_items]
+            for k in fields
+        }, pa.schema({
+            'id': pa.utf8(),
+            'label': pa.utf8(),
+            'type': pa.utf8(),
+            'columnNames': pa.list_(pa.utf8()),
+            'columnTypes': pa.list_(pa.utf8())
+        }))
+        outputs.set_value('selectedItems', selected_items)
+
 
 # Value of any column in the mapping table
 MappingItemStruct = pa.struct([
@@ -77,6 +142,7 @@ def build_table_from_mapping(
 
 
 class NetworkAnalysisDataMappingModule(KiaraModule):
+    _module_type_name = 'data_mapping'
 
     def create_input_schema(self) -> Mapping[str, ValueSchema]:
         return {
@@ -105,7 +171,7 @@ class NetworkAnalysisDataMappingModule(KiaraModule):
             )
         }
 
-    def process(self, inputs: StepInputs, outputs: StepOutputs) -> None:
+    def process(self, inputs: ValueSet, outputs: ValueSet) -> None:
         nodes_mapping_table = inputs.get_value_data(
             'nodesMappingTable') or pa.Table.from_pydict({})
         nodes = build_table_from_mapping(
