+++ lumy_middleware/kiara_modules/pipelines/__init__.py
@@ -2,7 +2,7 @@ from kiara import (KiaraEntryPointItem, find_pipeline_base_path_for_module)
 
 pipelines: KiaraEntryPointItem = (
     find_pipeline_base_path_for_module,
-    ["lumy_middleware.dev.pipelines"],
+    ["lumy_middleware.kiara_modules.pipelines"],
 )
 
 KIARA_METADATA = {"tags": ["pipeline"], "labels": {"pipeline": "true"}}
