+++ lumy_middleware/kiara_modules/pipelines/createTableFromFile.json
@@ -1,13 +1,14 @@
 {
-  "$schema": "https://dharpa.org/kiara/development/entities/modules/PipelineModuleConfig.json",
+  "$schema": "https://raw.githubusercontent.com/DHARPA-Project/kiara/gh-pages/latest/development/entities/modules/PipelineConfig.json",
   "module_type_name": "table_from_file",
+  "doc": "Create a table from a CSV file.",
   "steps": [
     {
-      "module_type": "lumydev.table.local_file",
+      "module_type": "lumy.table.local_file",
       "step_id": "read_file"
     },
     {
-      "module_type": "lumydev.table.create_from_file",
+      "module_type": "lumy.table.create_from_file",
       "step_id": "create_table_from_file",
       "input_links": {
         "file": "read_file.file"
