+++ lumy_middleware/notes/README.md
@@ -1,0 +1,3 @@
+# Notes
+
+Contains in memory notes backend implementation. Should be removed once notes taking is supported in Kiara.
